refactor(web): simplify filter modal — self-resolving ChipFacet, shared FacetHeader
- ChipFacet resolves its own options from the registry by param; drop 7 copy-paste
  FACETS.find(...).options consts and the options prop from every call site.
- Extract FacetHeader (label + Exclude + Clear) shared by ChipFacet and the
  Specialization pane, removing the duplicated header markup.
- FilterSummary resolves Location chip labels via the shared valueLabel (drops the
  direct REGION_LABELS/countryLabel imports).
No behavior change; svelte-check clean, verified in-app.
